Static json renames
