+++ packages/types/src/Metadata/v0/static-substrate.json
