+++ packages/types/src/Metadata/v6/Metadata.spec.ts
@@ -2,8 +2,8 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import polkadotJson from './latest.polkadot.v6.json';
-import substrateJson from './latest.substrate.v6.json';
+import polkadotJson from './static-polkadot.json';
+import substrateJson from './static-substrate.json';
 import polkadotData from './static.polkadot';
 import substrateData from './static';
 import { decodeLatestSubstrate, defaultValues, toV7 } from '../util/testUtil';
