+++ packages/types/src/Metadata/v6/static-substrate.json
