+++ packages/types/src/Metadata/v7/Metadata.spec.ts
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import substrateJson from './latest.substrate.v7.json';
+import substrateJson from './static-substrate.json';
 import substrateData from './static';
 import { decodeLatestSubstrate, defaultValues, toV7 } from '../util/testUtil';
