+++ packages/types/src/Metadata/v7/static-substrate.json
